+++ handlers/hello.go
@@ -9,9 +9,13 @@ import (
 	"github.com/kataras/iris/v12"
 )
 
+type HelloRes struct {
+	Message string `json:"message"`
+}
+
 func Hello(ctx iris.Context) {
 	translator := ctx.Values().Get(g.TranslatorKey).(i18n_interfaces.TranslatorI)
-	utils.SendJson(ctx, map[string]string{
-		"message": translator.HelloWorld() + " 🥳",
+	utils.SendJson(ctx, HelloRes{
+		Message: translator.HelloWorld() + " 🥳",
 	})
 }